backend/requirements.txt
# ═══════════════════════════════════════════════════════════════
# PLAYE Studio Pro v3.0 — STANDARD (16GB RAM, GTX 1060+ / 4GB VRAM)
#
# Что работает: Всё из Lite + face-enhance (RestoreFormer),
#   upscale (RealESRGAN), denoise (NAFNet), YOLO detection,
#   diffusers (inpaint, depth), SAM 2, face-id.
# Что НЕ работает: PaddleOCR, open3d, 3D reconstruction.
# ═══════════════════════════════════════════════════════════════

# Core
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.18
pydantic>=2.10.0
pydantic-settings>=2.7.0

# DB & Auth
sqlalchemy>=2.0.0
alembic>=1.13.0
bcrypt>=4.2.0
python-jose>=3.3.0

# Media
numpy>=1.26.0,<2.2.0
Pillow>=11.1.0
opencv-python-headless>=4.10.0
ffmpeg-python>=0.2.0
scipy>=1.14.0

# PyTorch (install.ps1 ставит CUDA-версию отдельно)
torch>=2.6.0
torchvision>=0.21.0
onnxruntime>=1.20.0

# AI Vision
easyocr>=1.7.0
ultralytics>=8.3.0
insightface>=0.7.3
diffusers>=0.32.0
transformers>=4.48.0
accelerate>=0.36.0
safetensors>=0.5.0

# Face restoration (install.ps1 ставит --no-deps)
# basicsr==1.4.2
# gfpgan==1.3.8
# realesrgan==0.3.0
# facexlib>=0.3.0

# Utils
requests>=2.32.0
tqdm>=4.67.0
lmdb>=1.5.0
addict>=2.4.0
yapf>=0.43.0
prometheus-fastapi-instrumentator>=7.0.0
